// src/app/model/flight/flight-list/flightSegments.ts
export class FlightSegments {
  segmentKey: string;
  depAirportCode: string;
  depAirportName: string;
  depCity: string;
  depCityName: string;
  depCountry: string;
  depTerminal: string;
  depDateTime: string;
  arrAirportCode: string;
  arrAirportName: string;
  arrCity: string;
  arrCityName: string;
  arrCountry: string;
  arrTerminal: string;
  arrDateTime: string;
  airline: string;
  airlineLogo: string;
  flightNumber: string;
  aircraft: string;
  marketingName: string;
  fareBasisCode: string;
  distance: string;
  duration: string;
}